feat: done cybersecurity modal page

+++ Assets/UI/Gameplay/Pages/Page1/Character/StatSlotUI.cs
@@ -8,7 +8,7 @@ public class StatSlotUI : MonoBehaviour
 
     [SerializeField] private TMP_Text Name;
     [SerializeField] private TMP_Text Value;
-    public void Setup(KeyValuePair<string, int> data)
+    public void Setup(KeyValuePair<string, float> data)
     {
         Name.text = data.Key;
         Value.text = data.Value.ToString();
